% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gofr.R
\name{Gofr}
\alias{Gofr}
\title{G(r): small sample correlation coefficient}
\usage{
Gofr(x, y)
}
\arguments{
\item{2}{numeric vectors}
}
\description{
Calculates an unbiased correlation coefficient for studies with small sample sizes, from Olkin and Pratt 1958
}
\references{
Olkin, I. and J. W. Pratt, 1958, Unbiased estimation of certain correlation coefficients, The Annals of Mathematical Statistics, 29(1): 201-211.
}

