% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gr.R
\name{Gr}
\alias{Gr}
\title{G(r): Small sample correlation coefficient}
\usage{
Gr(x, y)
}
\arguments{
\item{2}{columns of a data frame}
}
\description{
Calculates an unbiased correlation coefficient for studies with small sample sizes, from Olkin and Pratt 1958.
}
\examples{
Gr(stackloss$Air.Flow,stackloss$Acid.Conc.)
}
\references{
Olkin, I. and J. W. Pratt, 1958, Unbiased estimation of certain correlation coefficients, The Annals of Mathematical Statistics, 29(1): 201-211.
}

