% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedgesd.R
\name{hedges.d}
\alias{hedges.d}
\title{Hedges' d}
\usage{
hedges.d(x, y)
}
\arguments{
\item{2}{vectors}
}
\description{
Calculates Hedges' d effect size for studies with small sample sizes.  x and y represent observations from treatment and control groups of a controlled trial.
}
\references{
Hedges, L and I Olkin, 1985. Statistical Methods for Meta-Analysis. Academic Press, Orlando Fl. pp. 81, 104
}

