% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedgesd.R
\name{Hedgesd}
\alias{Hedgesd}
\title{Hedges' d}
\usage{
Hedgesd(x, y)
}
\arguments{
\item{2}{vectors}
}
\description{
Calculates Hedges' d effect size for studies with small sample sizes.  The two vectors represent observations of treatment and control groups in a controlled trial.
}
\examples{
Hedgesd(sleep[sleep$group==1,1],sleep[sleep$group==2,1])
}
\references{
Hedges, L and I Olkin, 1985. Statistical Methods for Meta-Analysis. Academic Press, Orlando Fl. pp. 81, 104
}

