% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRdelta.R
\name{RRdelta}
\alias{RRdelta}
\title{RRdelta}
\usage{
RRdelta(x, y)
}
\arguments{
\item{2}{vectors}
}
\description{
Calculates the unbiased response ratio and associated variance between two groups of a controlled trial using the delta method (Lajeunesse 2015).
}
\examples{
RRdelta(sleep[sleep$group==1,1],sleep[sleep$group==2,1])
}
\references{
Lajeunesse, M J.  2015.  Bias and correction for the log response ratio in ecological meta-analysis.  Ecology, 96(8): 2056-2063
}

