% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRdelta.R
\name{RRdelta}
\alias{RRdelta}
\title{RRdelta}
\usage{
RRdelta(x, y)
}
\arguments{
\item{2}{vectors}
}
\description{
Calculates the unbiased response ratio and variance for the effect a treatment in a controlled trial with two groups using the delta method (Lajeunesse 2015).
}
\references{
Lajeunesse, M J.  2015.  Bias and correction for the log response ratio in ecological meta-analysis.  Ecology, 96(8): 2056-2063
}

